refactor(phase-4): split zeventext into extensions/zevent/ package
Break the 1146-line zeventext.py god-extension into mixins under
extensions/zevent/:

- _common.py: config, URLs, intervals, event dates, StreamerInfo,
  split_streamer_list
- api.py: ApiMixin — streamer/planning caches, API validation,
  event-phase checks, donation total (Zevent + Streamlabs)
- streams.py: StreamsMixin — Twitch categorisation, viewer
  aggregation, follower-based sorting for offline online streamers
- embeds.py: EmbedsMixin — main/location/planning/top-donations
  embeds and the 6000-char size enforcement helper
- tasks.py: TasksMixin — the interval refresh task and milestone
  announcements
- commands.py: CommandsMixin — /zevent_finish

The glue Zevent class wires cache attributes, runs on_startup
to resolve the pinned message and auth Twitch, then starts the
refresh loop.
